Changing more about DCT, Added to the decoder
DCT wasn't working properly I disocvered, so through trial and error
I found one that does, I just need to adapt it now to work with what I
have.
Started making the decoder, it goes up to the inverse DCT right now.
I removed entropy coding from this, maybe after I get the DCT working
I'll
put it back in but as of right now it was not working so it's been
skipped.